Remove Singletons
